Updated standalone setup to register the custom factory.
Eclipse register RoboChartFactoryImplCustom using the plugin, but this
is not reflect when running the tools standalone. As a consequence the
constructed model does not use the custom implementations of classes
such as Variable. This leads to exceptions in the code generator and
validator.

The standalone setup class was updated to not only register the
EPackage, but also the modified EFactory.