Updated validator to deal with required variables of an operation.
Previously, the validator would check if the required variables of an
operation are satisfied by the controller that contains it. This is not
correct because the operation call is executed in the context of the
state machine, not the controller. I have changed the validator to check
that the required variables of an operation are available in the context
of the state machine in which the operation is called.

I have also updated the .classpath files to match Java 10.