Updated syntax for vector and matrix types to allow for general expressions in the size, rows and columns attributes. Added syntax for vector and matrix expressions.
Added validations to check that size, rows and columns are either
integer expressions or references to constants.

Updated type checker to deal with vector and matrix types without
directly instead of converting them to product types. Implemented type
checking of vecotr and matrix expressions, sum, subtraction,
multiplication, inversion and transposition of matrices and vector.

Added parsing tests for syntax and type checking of matrices and
vectors.

Extended printing services to deal with vector and matrix types.

Still need to extend graphical editor to deal with new expressions, in
particular, printing services of the graphical editor plugin.