Updated well-formedness rules.
- Weakened C4 to not consider operations:

Instead, this is catered by a weakened form, where: all operations
required by the controller's state machines, including the machines that
define operations, must be required or defined by the controller.

This is required to avoid requiring that state machines provide
transitively all data (variables) for all operations that are directly
or indirectly called.

Observations:

* I indent to explore the conditional application of full C4 in an
upcoming commit to ensure the current CSP generator is not used
inconsistently.

** There is a question regarding the use of OperationDefs directly,
without use of required interfaces. I believe this is because the scope
provider is not tight enough as currently such use yields broken CSP.

- Added STM8:

A state machine that requires an operation, including a machine that
defines another operation, must define, or require all variables
required by that operation. Note that the requirement in relation for
the variables provided by the controller is catered already by C3.

- Added STM9:

A state machine that requires an operation, including a machine that
defines another operation, must define all events defined in that
operation. I believe previously this was assumed to be implicitly
satisfied via the scope provider. However, there were cases were this
was not being checked.

- Added STM10:

A state machine that requires a clock, including a machine that defines
another operation, must define, or require all clocks required by that
operation. Note that STM7 prevents actual state machines (and not
operations) from requiring clocks.